Overhaul/refector/rewrite of MaRzAgent for new algorithm adjustments
Nux butchered my code.  Cleaned up several old methods.

We now track a node's success/fail rate via tracking indexes into
the memory list where that suffix occurs.  This redefines 'success' a
little because if a given sequence to try reaches the goal before it
reaches the suffix in the sequence we want to treat that as a success.

This code is not complete.  More updates needed and surely much debugging
too.

See journal notes for more details.