Fixed game-breaking bugs in NSMAgent.java
The Q-learning in this agent was not happening as it was supposed to.  I
added the right method calls and fixed some bugs in the learning update
method.

Along the way to finding this bug I added code to insert configuration
information automatically in the AIReport .csv file.  I modified the .csv
filename to include the datetime so as to guarantee a unique file each
run.

Finally I also implemented a partial unit test for NSMAgent.